refactor: SystemSettings 위치 및 패키지 변경
`SystemSettingsExtensions`를 `SystemSettings`로 이름을 변경하고 `ui.util` 패키지로 이동합니다.
